Move service::start
